# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for user management.",
        "title": "User API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/const/createConst": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "เพิ่มข้อมูลค่าคงที่",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Const"
                ],
                "summary": "เพิ่มข้อมูลค่าคงที่",
                "parameters": [
                    {
                        "type": "string",
                        "default": "th",
                        "description": "(en, th)",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": " request body ",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.ConfigConstant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/const/deleteConst/{group_id}/{const_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ลบข้อมูล ค่าคงที่",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Const"
                ],
                "summary": "ลบข้อมูล ค่าคงที่",
                "parameters": [
                    {
                        "type": "string",
                        "default": "th",
                        "description": "(en, th)",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "group id",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "const id",
                        "name": "const_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/const/findConst": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แสดงข้อมูลค่าคงที่แบบแบ่งหน้า",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Const"
                ],
                "summary": "แสดงข้อมูลค่าคงที่แบบแบ่งหน้า",
                "parameters": [
                    {
                        "type": "string",
                        "default": "th",
                        "description": "(en, th)",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "ไอดีค่าคงที่",
                        "name": "const_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ไอดีกลุ่มค่าคงที่",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "สถานะใช้งาน",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ชื่อค่าคงที่ EN",
                        "name": "name_en",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ชื่อค่าคงที่ TH",
                        "name": "name_th",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 1",
                        "name": "ref_value1",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 2",
                        "name": "ref_value2",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 3",
                        "name": "ref_value3",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ลำดับ",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.Pagination-arczed_internal_entities_models_ConfigConstant"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/const/findConstAll": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แสดงข้อมูลค่าคงที่ทั้งหมด",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Const"
                ],
                "summary": "แสดงข้อมูลค่าคงที่ทั้งหมด",
                "parameters": [
                    {
                        "type": "string",
                        "default": "th",
                        "description": "(en, th)",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "ไอดีค่าคงที่",
                        "name": "const_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ไอดีกลุ่มค่าคงที่",
                        "name": "group_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "สถานะใช้งาน",
                        "name": "is_active",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ชื่อค่าคงที่ EN",
                        "name": "name_en",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ชื่อค่าคงที่ TH",
                        "name": "name_th",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 1",
                        "name": "ref_value1",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 2",
                        "name": "ref_value2",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ค่าอ้างอิง 3",
                        "name": "ref_value3",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ลำดับ",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.ConfigConstantResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/const/updateConst": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แก้ไขข้อมูลค่าคงที่",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Const"
                ],
                "summary": "แก้ไขข้อมูลค่าคงที่",
                "parameters": [
                    {
                        "type": "string",
                        "default": "th",
                        "description": "(en, th)",
                        "name": "Accept-Language",
                        "in": "header"
                    },
                    {
                        "description": " request body ",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.ConfigConstant"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/arczed_internal_entities_schemas.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "arczed_internal_entities_schemas.ConfigConstant": {
            "type": "object",
            "properties": {
                "const_id": {
                    "description": "ไอดีค่าคงที่",
                    "type": "string"
                },
                "group_id": {
                    "description": "ไอดีกลุ่มค่าคงที่",
                    "type": "string"
                },
                "is_active": {
                    "description": "สถานะใช้งาน",
                    "type": "string"
                },
                "name_en": {
                    "description": "ชื่อค่าคงที่ EN",
                    "type": "string"
                },
                "name_th": {
                    "description": "ชื่อค่าคงที่ TH",
                    "type": "string"
                },
                "ref_value1": {
                    "description": "ค่าอ้างอิง 1",
                    "type": "string"
                },
                "ref_value2": {
                    "description": "ค่าอ้างอิง 2",
                    "type": "string"
                },
                "ref_value3": {
                    "description": "ค่าอ้างอิง 3",
                    "type": "string"
                },
                "sort": {
                    "description": "ลำดับ",
                    "type": "integer"
                }
            }
        },
        "arczed_internal_entities_schemas.ConfigConstantResp": {
            "type": "object",
            "properties": {
                "const_id": {
                    "description": "ไอดีค่าคงที่",
                    "type": "string"
                },
                "created_at": {
                    "description": "วันที่สร้าง",
                    "type": "string"
                },
                "created_user": {
                    "description": "ผู้สร้าง",
                    "type": "string"
                },
                "deleted_at": {
                    "description": "วันเวลาที่ลบ",
                    "allOf": [
                        {
                            "$ref": "#/definitions/gorm.DeletedAt"
                        }
                    ]
                },
                "deleted_user": {
                    "description": "ผุ้ลบ",
                    "type": "string"
                },
                "group_id": {
                    "description": "ไอดีกลุ่มค่าคงที่",
                    "type": "string"
                },
                "is_active": {
                    "description": "สถานะใช้งาน",
                    "type": "integer"
                },
                "is_active_txt": {
                    "description": "สถานะใช้งาน",
                    "type": "string"
                },
                "name_en": {
                    "description": "ชื่อค่าคงที่ EN",
                    "type": "string"
                },
                "name_th": {
                    "description": "ชื่อค่าคงที่ TH",
                    "type": "string"
                },
                "ref_value1": {
                    "description": "ค่าอ้างอิง 1",
                    "type": "string"
                },
                "ref_value2": {
                    "description": "ค่าอ้างอิง 2",
                    "type": "string"
                },
                "ref_value3": {
                    "description": "ค่าอ้างอิง 3",
                    "type": "string"
                },
                "sort": {
                    "description": "ลำดับ",
                    "type": "integer"
                },
                "updated_at": {
                    "description": "วันเวลาที่อัพเดทล่าสุด",
                    "type": "string"
                },
                "updated_user": {
                    "description": "ผู้อัพเดทล่าสุด",
                    "type": "string"
                }
            }
        },
        "arczed_internal_entities_schemas.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "arczed_internal_entities_schemas.Pagination-arczed_internal_entities_models_ConfigConstant": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "rows": {
                    "description": "เพิ่ม swaggertype:\"array,object\" เพื่อให้ swagger init ได้",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "sort": {
                    "type": "string"
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_rows": {
                    "type": "integer"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "ใส่ค่า Bearer เว้นวรรคและตามด้วย TOKEN  ex(Bearer ?????????)"
        }
    }
}